requirements.txt
Django==1.7.1
django-contact-form==1.0
django-debug-toolbar==1.2.2
django-model-utils==2.2
psycopg2==2.5.4
selenium==2.43.0
sorl-thumbnail==11.12.1b
sqlparse==0.1.13
